requirements_cuda.txt
--index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://pypi.org/simple
torch==2.2.2
torchvision==0.17.2
transformers>=4.38.0
datasets>=2.18.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
jiwer>=3.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0
ultralytics>=8.0.0
PyYAML>=6.0
tqdm>=4.66.0
accelerate>=0.27.0